Feature/rttov interface updates (#324)
* first attempt to support lfric only
what have I forgotten?

* string literals didn't have length MAXVARLEN which was expected by
ufo_geovals_get_var. Now using varname in call which does

* fix string literal in getvar call

* compiles

now able to allocate memory for a fraction of the obs and process the
chunks (I think - not tested, or at least only with a
single observation)

BTs coming out of RTTOV appear to be being reported as negative? Will
investigate

Seems to be stopping at the TL/AD?

The biggest problem next is how to resolve the issue that as it stands
the jacobians must be stored for every single profile. It will
probably take a few lines of code and not being so lazy... will fix
tomorrow!

* could be that all the fields are the wrong way up

* commit work in progress to support lfric.

gets through to end but Jo in ginormous so something might be wrong

next need to re-add support for amsua test and test to see which way
up profile is

* fix a couple of typos (necessary for more than one profile)
remember to nullify profiles_k
try again

* now on levels not layers because that's what lfric passes us

* added channels and synced with CRTM
compiles but crashes before we get to RTTOV with no output

* emacs backup files. Don't want these

* consistent naming with CRTM and missed a header

* do the merge work that mergetool said it did!

* Reduce IO burden to handle 1e6 obs.

Only read coefs once now and store in public all_profs array imported
from the utils_mod.

A bunch of other stuff too (e.g. possibly better handling of bad obs
but need a more diverse test to test this capability fully).

Some sort of get_emissivity routine now which can be added to as
necessary.

Also requires a small tweak to RTTOV to handle resetting the profile
in the case of excessively low humidity (otherwise you end up with a
3GB err file when RTTOV is being verbose).

This commit builds and runs the one ob and the 1e6 ob case and returns some
output from the direct. It's giving NaNs for the 1e6 case but that
could be for lots of reasons.

* code tidy

delete loads of debugging junk. Might reimplement with proper file
names and fckit if necessary.

Removed outputting all the BTs but you might want them.

* Revert "code tidy"

This reverts commit 9a0227d71136819d30c3b1a603d22ee66d0a917d.

* Revert "Reduce IO burden to handle 1e6 obs."

This reverts commit b200898286c3e164da9c98ecbc3fea060b87967e.

* removed write statement

* intermediate refactor commit

* fixed issue with geovals_read; renamed geovals nobs -> nlocs. need to replace aircraft files (remove an obs outside of time window)

* added a file with aircraft geovals (all obs within timewindow)

* fix bug in ufo_geovals_delete

* fixed bug in ufo_geovals_delete (deallocating variables)

* removed comment

* removing print statements from DR's code

* fixed GeoVaLs sections of yaml files

* Fixed the names to netcdf cf for radiance.

* update of RTTOV

* Moved filters to sub-directory

* cleanup test config files

* parts of marine build/test disabled if GSW not found

* make config use consistent with updated obsconfig

* changes associated with obsconfig_cleanup

* use pressure instead of log(p) in atmvertinterp

* updating code to work with RTTOV.

* use air_pressure in Pa in crtm, rttov, atmvertinterp

* change angle metadata to netcdf-cf for rttov

* working on ufo rttov tests with aquaplanet geovals.

* more tests - the problem is that the meta data fields are not being populated

* tidy up.

* responding to reviewers comments and tidying up.

* updated test

* removed print statement in geovals

* update geovals files to use air_pressure in Pa

* remove LittleEndian as it is not being used.

* also remove prsl use in the analytical initial conditions (#328)

* also remove prsl use in the analytical initial conditions

* remove var_prsl

* allow for azimuth and solar angles and rescale pressure in geovals to Pa

* add azimuth and solar angles

* remove include to obsspace_mod

* air_pressure unit changed to Pa

* all operators use pressure-Pa

* slight adjustment of rmsequiv due to pressure unit change