Feature/rttov mar20 update (#1284)
* WIP commit
see where the errors are (everywhere) - not expecting anything to work

* removed TLAD for a moment whilst I'm getting new look direct sorted
New look RTTOV interface first go compiles but not tested

* done TLAD
it compiles
but no tests run yet

* 'passes' first comparison with CRTM at expected level

* hofxdiags probably working now for direct
loads of debug stuff still in there

* removed debugging noise

* TL and AD finally updated with q, clw, various surface variables. All
of them are required at the moment which will need fixing.

Various satrad hacks which I would prefer weren't in here but they're
on a switch.

Tidied up a bunch of stuff but there's still debugging code (probably
commented out).

* correctly commenting out old vv array

* Tidied up code
Fixed miscellaneous errors
Fixed TLAD interface (added surface variables to output vector)
Added some RTTOV specific conversions that probably need to be sourced
from somewhere else (e.g. Pa_to_hPa)
Added RTTOV required variables to ufo_variables_mod
Updated the basic amsua_rttov test and will follow up with something
more comprehensive soon

* code tidy (lowercase keywords, remove commented code, (re)move
comments)

* update tests
  - remove old test of interface v1
  - add more modern test with OPS fields

* fix atms/amsua typo. Thanks Stu

* Fix coding norms by deleting commented out code

* Changes in response to PR #1284.

Moved Ops_Qsat to new routine ufo_utils_mod.F90
This isn't compiling because of some problem with c_ptr but I wasn't
aware I'd changed anything so I'll merge ufo again

* whoops. The changes suggested by reviewers as part of PR #1283

* clone rrtov in aws codebuild

* bugfix

* many fixes to get things working again (tests all pass)

I think the potential OpenMP issue is reolved now. Removing the =>
null() in the declaration caused problems and highlighted an issue
where we were trying to use deallocated memory in the TLAD. So that
required the allocation code to be rewritten. It looks pretty
miserable now and should be revisited ASAP.

renamed RTProf to RTprof_K in the TLAD code to make it clear where we are

Refer to specific constants in ufo_constants_mod
made intent of parse_hofxdiags a bit clearer (explicitly sets
jacobian_needed)

* replace fckit_log%info calls with abor1_ftn calls for terminal errors

* add the geovals which I had omitted
add dependence of tests on the get ioda data task
change PROGRAM_NAME to ROUTINE_NAME and use it to trace where errors are coming from

* omitted the symlink to the geovals for the ctest. Fixed here.

* trigger pipeline

* Fix initialisation of emissivity
Change Fastem defaults to RTTOV defaults even though they are not used
because they're causing a fatal error in RTTOV which is causing a
whole batch of obs to be lost

* re-instate fastem initialisation as that was a red-herring. It's
useful to have them initialised to 0 because we're not supposed to be
using them so it's a good error to catch.

The problem was a missing check on skip profiles in the emissivity
routine. I have taken the opportunity to tidy up some of the
initialisations too.

I have also done a bit more tidying and added

a 'Debug' switch (default F) to the obs operator yaml (to control the
noise) and

a 'max_channels_per_batch' switch to control the batching issue
(default still 10000). I suspect I will replace this with a max_prof
because setting the number of channels to less than one profiles worth
gives a segfault (more is fine). But max_prof is not ideal either
because the point is to not stuff RTTOV too full of channels. A work
in progress.

rttov ctests still pass

* fix problem with skip_profiles at the end of the array

* fix logic in init_emissivity: emissivity must be initialised for every
RTTOV channel. Skipped profiles don't count and shouldn't be skipped
in calcemis or emissivity. This was causing problems.

This has now been tested with the test cases and Stu's test for a
number of tiny, small, medium, big and huge chunk sizes and runs.

fix keyword capitalisation
make nchan vs nchans consistent (all nchan)

* Bumping automatic build system

Co-authored-by: Maryam Abdi-Oskouei <maryamao@ucar.edu>
Co-authored-by: Ryan Honeyager <honeyage@ucar.edu>