obserr_/flags_, remove hardcoded groups, fix tests
Use obserr_, flags_ instead of get/putObservation because the flag and
error data is already allocated there. This means that we can remove the
hard-coded retrievals from `ObsError` and `QCFlags`. This also means
that QC and error data is now consistent with the missing values and so
some further changes to the tests that use these transforms are required.